### 2022-05-26 #4428 @editorialbot
[PRE REVIEW]: special: A Python package for the spectral characterization of directly imaged low-mass companions
**Submitting author:** @VChristiaens (<a href="http://orcid.org/0000-0002-0101-8814">Valentin Christiaens</a>)
**Repository:** https://github.com/VChristiaens/special
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2
