### 2022-06-06 #4456 @editorialbot
[REVIEW]: special: A Python package for the spectral characterization of directly imaged low-mass companions
**Submitting author:** @VChristiaens (<a href="http://orcid.org/0000-0002-0101-8814">Valentin Christiaens</a>)
**Repository:** https://github.com/VChristiaens/special
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.4
