### 2022-06-06 #4456 @dfm
Jonathan will do most of his review after some travel that he's doing for the next few weeks and the bot will send a ping then.

@jgagneastro: Before you leave, can you comment `@editorialbot generate my checklist` on this thread and go through the conflict of interest just so it's there?

@QuentChangeat: You're welcome / encouraged to start your review earlier! This will make things even smoother for when @jgagneastro returns.
