### 2022-06-07 #4456 @jgagneastro
A reminder for myself: (1) Should add link to tutorials in README directly (2) should replace the path "../datasets/CrA-9_2019_IFS_pca_annulus.fits" with os.path.dirname(__file__)+"/datasets/CrA-9_2019_IFS_pca_annulus.fits" such that the tutorial works regardless of where we are cd'ed
