### 2022-06-10 edf3539
Modified make docs behaviour (delete special.rst file first)
