### 2022-06-13 f6e6017
Improved documentation: reference hyperlinks solved + nice formatting
