+++ special/mcmc_sampling.py
@@ -8,15 +8,37 @@ estimation.
    | Foreman-Mackey et al. 2013
    | **emcee: The MCMC Hammer**
    | *PASP, Volume 125, Issue 925, p. 306*
-   | `https://arxiv.org/abs/astro-ph/1202.3665
-     <https://arxiv.org/abs/astro-ph/1202.3665>`_
+   | `https://arxiv.org/abs/1202.3665
+     <https://arxiv.org/abs/1202.3665>`_
      
 .. [FOR19]
    | Foreman-Mackey et al. 2019
    | **emcee v3: A Python ensemble sampling toolkit for affine-invariant MCMC**
    | *JOSS, Volume 4, Issue 43, p. 1864*
-   | `https://arxiv.org/abs/astro-ph/1911.07688
-     <https://arxiv.org/abs/astro-ph/1911.07688>`_
+   | `https://arxiv.org/abs/1911.07688
+     <https://arxiv.org/abs/1911.07688>`_
+     
+.. [GOO10]
+   | Goodman & Weare 2010
+   | **Ensemble samplers with affine invariance**
+   | *Comm. App. Math. Comp. Sci., Vol. 5, Issue 1, pp. 65-80.*
+   | `https://ui.adsabs.harvard.edu/abs/2010CAMCS...5...65G
+     <https://ui.adsabs.harvard.edu/abs/2010CAMCS...5...65G>`_
+
+.. [GRE16]
+   | Greco & Brandt 2016
+   | **The Measurement, Treatment, and Impact of Spectral Covariance and 
+     Bayesian Priors in Integral-field Spectroscopy of Exoplanets**
+   | *The Astrophysical Journal, Volume 833, Issue 1, p. 134*
+   | `https://arxiv.org/abs/1602.00691
+     <https://arxiv.org/abs/1602.00691>`_     
+
+.. [OLO16]
+   | Olofsson et al. 2016
+   | **Azimuthal asymmetries in the debris disk around HD 61005. A massive collision of planetesimals?**
+   | *Astronomy & Astrophysics, Volume 591, p. 108*
+   | `https://arxiv.org/abs/1601.07861
+     <https://arxiv.org/abs/1601.07861>`_
      
 """
 
@@ -59,15 +81,15 @@ def lnprior(params, labels, bounds, priors=None):
     params: tuple
         The model parameters.
     labels: Tuple of strings
-        Tuple of labels in the same order as initial_state, that is:
-        - first all parameters related to loaded models (e.g. 'Teff', 'logg')
-        - then the planet photometric radius 'R', in Jupiter radius
-        - (optionally) the flux of emission lines (labels should match those in
-        the em_lines dictionary), in units of the model spectrum (times mu)
-        - (optionally) the optical extinction 'Av', in mag
-        - (optionally) the ratio of total to selective optical extinction 'Rv'
-        - (optionally) 'Tbb1', 'Rbb1', 'Tbb2', 'Rbb2', etc. for each extra bb
-        contribution.
+        Tuple of labels in the same order as initial_state:
+            - first all parameters related to loaded models (e.g. 'Teff', 'logg')
+            - then the planet photometric radius 'R', in Jupiter radius
+            - (optionally) the flux of emission lines (labels should match those \
+            in the em_lines dictionary), in units of the model spectrum (times mu)
+            - (optionally) the optical extinction 'Av', in mag
+            - (optionally) the ratio of total to selective optical extinction 'Rv'
+            - (optionally) 'Tbb1', 'Rbb1', 'Tbb2', 'Rbb2', etc. for each extra bb \
+            contribution. 
     bounds: dictionary
         Each entry should be associated with a tuple corresponding to lower and 
         upper bounds respectively. Bounds should be provided for ALL model
[…]
