### 2022-06-13 #4456 @VChristiaens
Thanks for taking the lead as editor @jfm, and thanks @jgagneastro and @QuentChangeat for accepting to review this submission!
I have implemented your first suggestions in my latest push @jgagneastro. In fact the tutorial links were already in the README, but having another look at it made me realize I had forgotten to mention my latest addition to `special` - that is compatibility with the UltraNest sampler. Now this is added to the README and documentation. Your suggestion of changing the path in the tutorial is now also implemented. Thanks.
