### 2022-06-13 cbc97f7
Documentation: improved aesthetics and removed duplicate links
