+++ special/spec_corr.py
@@ -16,7 +16,7 @@ from scipy.optimize import curve_fit
 def spectral_correlation(array, awidth=2, r_in=1, r_out=None, pl_xy=None,
                          mask_r=4, fwhm=4, sp_fwhm_guess=3, full_output=False):
     """ Computes the spectral correlation between (post-processed) IFS frames, 
-    as a function of radius, implemented as Eq. 7 of [GRE16]. This is a crucial 
+    as a function of radius, implemented as Eq. 7 of [GRE16]_. This is a crucial 
     step for an unbiased fit of a measured IFS spectrum to either synthetic or 
     template spectra.
     
