### 2022-06-13 12b4f98
Documentation: added small description to missing docs
