### 2022-07-02 b37d2eb
Minor bug fix in resample_model() function - lbda_min replaced by lbda_max
