### 2022-07-08 #4456 @QuentChangeat
Hi Valentin, I have started the review, special looks good. I have a few comments / suggestions:

1) It could be nice to have a higher level description, similar to the one in the accompanying paper, also at the beginning of the documentation.

2) The sources of the data does not seem to be stated in the documentation. Also, are the two provided files for CrA-9 in a standard format for the community?

3) It would be good to give more details/descriptions about the two files and how they were obtained. For instance, how are they obtained. Was the spectrum extracted from the annulus file via some external process?

4) Is grid fitting a standard procedure in DI? For spectroscopy of transiting planets, I believe the practice is to have the model computed live (see for instance the TauREx code: https://taurex3-public.readthedocs.io/en/latest/).
