### 2022-07-19 #4456 @VChristiaens
Hi @QuentChangeat, thank you for your feedback!
1. I have now updated the home page of the documentation as suggested;
2. I have now made the reference to the paper in which the data are published clearer, including a hyperlink, at the location where the data are loaded (it was further down in the tutorial) - FITS files are standard format for both reduced images and measured spectra in the direct imaging community;
3. I have now added a few more sentences at the beginning of section 1 of the notebook, to better explain how the spectrum was measured.
4. Most models used in the direct imaging community (e.g. BT-SETTL, DRIFT-PHOENIX, SONORA, EXOREM, ATMO) typically take a long time to compute as they rely on complex cloud micro-physics. The codes producing these models are also typically not made public. So interpolation in pre-computed grids of models has been the most common fitting method in the community for the past 20+ years. My impression is that it is still by far the standard in the community. It is true that models computed live have started to emerge. However, I think both complex cloud models and time-efficient models computed on the fly (which bear the risk of making oversimplifications) are still highly complementary at this stage.
