### 2022-07-21 #4456 @jgagneastro
Hey Valentin, sorry for the delay. I got back from vacation last week, but I was slow to get into the dozens of things I have gotten myself into before vacations because I came back with COVID. I'm now 100% and finally getting through the stuff I need to do.

I have another comment about this specific point " A statement of need: Do the authors clearly state what problems the software is designed to solve **and who the target audience is?**"

I think it would be useful to add a short paragraph in https://special.readthedocs.io/en/latest/features.html to say more explicitly that this package is specialized for the study of directly imaged exoplanets.

About this, can special be also used more widely to analyze the spectra and/or photometry of isolated brown dwarfs ? If so, it would be worth also mentioning it there. I'm sure many would like to use some of special's features even on isolated brown dwarfs if that doesn't cause any issue with the package.
