### 2022-07-21 b546a92
Clearer description of features in special, for identification of target audience
