+++ docs/features.rst
@@ -1,11 +1,15 @@
 What makes it `special`?
 ------------------------
 
-This package provides the following tools for the analysis of measured spectra:
+The ``special`` package provides a number of specific tools for the analysis of the spectrum of any (sub)stellar object. That being said, the main routines in the package (e.g. Bayesian retrieval of model parameters though MCMC or nested samplers, or best-fit template search) can also be applied to the spectrum of any object, provided a relevant grid of models or library of templates for the fit. 
 
-* calculation of the spectral correlation between channels of an IFS datacube;
+The format of the input spectrum is very flexible as it can consist of multi-band photometry, low-resolution or medium-resolution spectrum (or a combination thereof), obtained through direct imaging or not.
+
+The main available features of the package are listed below:
+
+* calculation of the spectral correlation between channels of an IFS datacube (relevant to directly imaged companions with an IFS, where the stellar halo had to be subtracted leaving residual speckle noise);
 * calculation of empirical spectral indices for MLT-dwarfs;
-* fitting of input spectra to different grids of models, including additional parameters such as (extra) black body component(s) and extinction;
+* fitting of input spectra to either grids of models or a blackbody, including additional parameters such as (extra) black body component(s), extinction and total-to-selective extinction ratio;
 * using either MCMC (`emcee <https://ui.adsabs.harvard.edu/abs/2013PASP..125..306F/abstract>`_) or nested (`nestle <http://github.com/kbarbary/nestle>`_ or `UltraNest <https://johannesbuchner.github.io/UltraNest/>`_) samplers to infer posterior distributions on spectral model parameters;
 * searching for the best-fit template spectrum within a given template library, with up to two free parameters (flux scaling and relative extinction).
 
@@ -14,11 +18,11 @@ The MCMC and nested sampler routines have been adapted to:
 
 * be flexible, as they are usable on any grid of models provided by the user (along with a snippet function specifying the format of the input);
 * sample the effect of (additional) blackbody components;
-* sample the effect of extinction; 
-* sample different extinction laws than ISM (parametrised using RV);
+* sample the effect of extinction (AV); 
+* sample different extinction laws than ISM (parametrised using the total-to-selective extinction ratio RV);
 * accept either uniform or Gaussian priors for each model parameter;
 * accept a prior on the mass of the object (if surface gravity is one of the model parameters, and for the MCMC sampler only);
-* consider convolution with the spectral PSF, photometric filters transmission and/or resampling of the model for consistency with the input spectrum.
+* consider convolution with the line spread function, photometric filters transmission and/or resampling of the model for consistency with the input spectrum - in particular convolution and resampling are done in two consecutive steps, and multiple resolving powers can be provided as input.
 * use a log-likelihood expression that can include i) spectral correlation between measurements of adjacent channels of a given instrument, and ii) additional weights that are proportional to the relative spectral bandwidth of each measurement, in case these are obtained from different instruments (e.g. photometry+spectroscopy).
 
-More details are available in `Christiaens et al. (2021) <https://ui.adsabs.harvard.edu/abs/2021MNRAS.502.6117C/abstract>`_ (note it was originally implemented as ``specfit``, a former module of the ``VIP`` package, before undergoing significant expansion).
+``special`` was originally implemented as ``specfit``, a former module of the ``VIP`` package, before undergoing significant expansion. It was first presented in `Christiaens et al. (2021) <https://ui.adsabs.harvard.edu/abs/2021MNRAS.502.6117C/abstract>`_ . More details will be available in an upcoming publication (Christiaens et al., subm. to JOSS).
