### 2022-07-21 #4456 @jgagneastro
Regarding "State of the field: Do the authors describe how this software compares to other commonly-used packages?", I think it would be relevant to discuss similarities/differences with Adam Burgasser's splat python package in the docs.
