### 2022-07-21 #4456 @jgagneastro
Regarding "References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?" The references seem appropriate to me in the article proofs, but the pandoc.org manual seems to require having citations between square brackets ? This seems odd to me, @dfm is this really required ?
