### 2022-07-21 #4456 @jgagneastro
Valentin, I am a bit worried about the use of Allers et al. (2007) Na I index with a spectral resolution much lower than SpeX-prism. Was this tested for a set of field and low-gravity dwarfs in the literature and I'm just not aware of it ? I could see this causing a lot of low-gravity false-positives.
