### 2022-07-21 27a9098
Added comparison to splat in paper draft
