+++ paper.bib
@@ -1,12 +1,34 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Valentin Christiaens at 2022-05-24 18:24:35 +0200 
+%% Created for Valentin Christiaens at 2022-07-21 21:58:36 +0200 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
 
 
+
+@inproceedings{Burgasser:2017,
+	abstract = {I describe our team's development of the SpeX Prism Library Analysis Toolkit (SPLAT), a combined spectral data repository for over 2500 low-resolution spectra of very low-mass dwarfs (late M, L and T dwarfs), and a Python-based analysis toolkit. SPLAT was constructed through a collaborative, student-centred, research-based model with high school, undergraduate and graduate students and regional K-12 science teachers. The toolkit enables spectral index analysis, classification, spectrophotometry, atmosphere model comparisons, population synthesis, and other analyses. I summarise the current components of this code, sample applications, and future development plans.},
+	adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+	adsurl = {https://ui.adsabs.harvard.edu/abs/2017ASInC..14....7B},
+	archiveprefix = {arXiv},
+	author = {{Burgasser}, A.~J. and {Splat Development Team}},
+	booktitle = {Astronomical Society of India Conference Series},
+	date-added = {2022-07-21 21:58:31 +0200},
+	date-modified = {2022-07-21 21:58:35 +0200},
+	eprint = {1707.00062},
+	keywords = {astronomical data bases: miscellaneous, stars: low-mass, brown dwarfs, techniques: spectroscopic, Astrophysics - Solar and Stellar Astrophysics},
+	month = jan,
+	pages = {7-12},
+	primaryclass = {astro-ph.SR},
+	series = {Astronomical Society of India Conference Series},
+	title = {{The SpeX Prism Library Analysis Toolkit (SPLAT): A Data Curation Model}},
+	volume = {14},
+	year = 2017,
[…]
