+++ README.rst
@@ -13,22 +13,34 @@
     :target: https://github.com/VChristiaens/special/blob/master/LICENSE
 
 
-``special`` is a package for the SPEctral Characterization of directly ImAged Low-mass companions. While some tools are specific to the characterisation of low-mass (M, L, T) dwarfs down to giant planets, ``special`` can also be used in a more general way for the characterisation of any object with a measured spectrum, provided an input model grid.
+``special`` is a package for the SPEctral Characterization of directly ImAged Low-mass companions. While some tools are specific to the characterisation of low-mass (M, L, T) dwarfs down to giant planets, ``special`` can also be used in a more general way for the characterisation of any object with a measured spectrum, provided an input model grid, regardless of the observational method used to obtain the spectrum (direct imaging or not) and regardless of the format of the spectra (multi-band photometry, low-resolution or medium-resolution spectrum, or a combination thereof).
 
 This package provides the following tools for the analysis of measured spectra:
 
-* calculation of the spectral correlation between channels of an IFS datacube;
+* calculation of the spectral correlation between channels of an IFS datacube (relevant to directly imaged companions with an IFS, where the uncertainty reflects spectrally correlated residual speckle noise);
 * calculation of empirical spectral indices for MLT-dwarfs;
-* fitting of input spectra to different grids of models, including additional parameters such as (extra) black body component(s) and extinction;
-* using either MCMC (`emcee <https://ui.adsabs.harvard.edu/abs/2013PASP..125..306F/abstract>`_) or nested (`nestle <http://github.com/kbarbary/nestle>`_ or `UltraNest <https://johannesbuchner.github.io/UltraNest/>`_) samplers to infer posterior distributions on spectral model parameters;
-* searching for the best-fit template spectrum within a given template library, with up to two free parameters (flux scaling and relative extinction).
+* fitting of input spectra to either photo-/atmospheric model grids or a blackbody model, including additional parameters such as (extra) black body component(s), extinction and total-to-selective extinction ratio;
+* using either MCMC (`emcee <https://ui.adsabs.harvard.edu/abs/2013PASP..125..306F/abstract>`_) or nested (`nestle <http://github.com/kbarbary/nestle>`_ or `UltraNest <https://johannesbuchner.github.io/UltraNest/>`_) samplers to infer posterior distributions on spectral model parameters in a Bayesian framework;
+* searching for the best-fit template spectrum within a given template library, with up to two free parameters (relative flux scaling and extinction).
 
-More details are available in `Christiaens et al. (2021) <https://ui.adsabs.harvard.edu/abs/2021MNRAS.502.6117C/abstract>`_ (note it was originally implemented as ``specfit``, a former module of the ``VIP`` package, before undergoing significant expansion).
+
+The MCMC and nested sampler routines have been adapted to:
+
+* be flexible, as they are usable on any grid of models provided by the user (along with a snippet function specifying how to read the format of the input files);
+* sample the effect of (additional) blackbody components;
+* sample the effect of extinction (AV); 
+* sample different extinction laws than ISM (parametrised using the total-to-selective extinction ratio RV);
+* sample a list of potential emission lines;
+* accept either uniform or Gaussian priors for each model parameter;
+* accept a prior on the mass of the object (if surface gravity is one of the model parameters, and for the MCMC sampler only);
+* consider convolution with the line spread function, photometric filters transmission and/or resampling of the model for consistency with the input spectrum - in particular convolution and resampling are done in two consecutive steps, and multiple resolving powers can be provided as input;
[…]
