### 2022-07-21 #4456 @VChristiaens
Thanks for the suggestion, I have now updated the paper draft including a comparison to ``splat``. I am a bit reluctant to add it on the readthedocs documentation though, mainly for two reasons:
(i) although common to find such comparison in a paper introduction to motivate new work, it feels somewhat too aggressive (to me) towards the other package and its maintainer when on a documentation page;
(ii) some of the features that are currently unique to ``special`` may eventually get implemented in ``splat`` without me knowing it as I do not check it regularly (not at all in fact, and would not want to remind myself to do that to update my documentation). I think a comparison at a given point in time is fine, such as in a paper.
Let me know if you still find it very relevant to include the comparison in the docs as well though.

> Regarding "State of the field: Do the authors describe how this software compares to other commonly-used packages?", I think it would be relevant to discuss similarities/differences with Adam Burgasser's splat python package in the docs.
