### 2022-07-21 #4456 @VChristiaens
You're right this is risky (I do not know either whether these were tested at low spectral resolution). At the moment, a warning is raised when the routines calculating spectral or gravity indices are called with a low-resolution input spectrum. I can adapt the behaviour to  raise an error instead, when the wavelength sampling of the input spectrum is sparser than a given threshold, if you deem it more appropriate (and remove the example from the tutorial). What do you think?

> Valentin, I am a bit worried about the use of Allers et al. (2007) Na I index with a spectral resolution much lower than SpeX-prism. Was this tested for a set of field and low-gravity dwarfs in the literature and I'm just not aware of it ? I could see this causing a lot of low-gravity false-positives.
