### 2022-07-26 ae79a84
Improved description for resample_model routine
