+++ special/model_resampling.py
@@ -561,8 +561,18 @@ def resample_model(lbda_obs, lbda_mod, spec_mod, dlbda_obs=None,
                    instru_res=None, instru_idx=None, filter_reader=None,
                    no_constraint=False, verbose=False):
     """
-    Convolve, interpolate and resample a model spectrum to match observed 
+    Convolve or interpolate, and resample, a model spectrum to match observed 
     spectrum.
+    
+    Note
+    ----
+    For models with large number of points (e.g. >1e6) and several resolving
+    powers are requested for convolution, this routine can be very slow. It can 
+    be useful to first bin down the model spectrum to ~50x the sampling of the 
+    observed spectrum before performing the convolutions corresponding to 
+    different resolving powers. Rebinning without convolution can be performed 
+    with this same function, by providing appropriate values only for the first 
+    3 parameters.
 
     Parameters
     ----------
@@ -589,25 +599,25 @@ def resample_model(lbda_obs, lbda_mod, spec_mod, dlbda_obs=None,
         measurements), interpolation (the opposite), or convolution by the 
         transmission curve of a photometric filter. If not provided, it will be 
         inferred from the difference between consecutive lbda_obs points (i.e. 
-        inaccurate for a combined spectrum). It must be provided IF one wants to 
-        weigh each measurement based on the spectral resolution of each 
-        instrument (as in [OLO16]_), through the ``use_weights`` argument.
+        inaccurate for a combined spectrum obtained with different instruments).
     instru_res : float or list of floats/strings, optional
         The mean instrumental resolving power(s) OR filter names. This is 
-        used to convolve the model spectrum. If several instruments are used, 
-        provide a list of resolving power values / filter names, one for 
-        each instrument used.
+        used to convolve the model spectrum. If several instruments/resolving
+        powere are to be considered, provide a list of resolving power values 
+        or filter names.
     instru_idx: numpy 1d array, optional
-        1d array containing an index representing each instrument used 
-        to obtain the spectrum, label them from 0 to the number of instruments 
-        (:math:`n_{ins}`). Zero for points that don't correspond to any of the 
-        ``instru_res`` values provided, and i in :math:`[1,n_{ins}]` for points
-        associated to instru_res[i-1]. This parameter must be provided if the 
-        spectrum consists of points obtained with different instruments.
+        1d array containing an index representing each instrument/resolving 
+        power used to obtain the spectrum. Label them from 0 to the number of 
+        instruments (:math:`n_{ins}`): zero for points that don't correspond to 
+        any of the ``instru_res`` values provided, and i in :math:`[1,n_{ins}]` 
+        for points associated to instru_res[i-1]. This parameter must be 
+        provided if the spectrum consists of points obtained with different 
+        instruments/resolving powers.
     filter_reader: python routine, optional
         External routine that reads a filter file and returns a 2D numpy array, 
         where the first column corresponds to wavelengths, and the second 
-        contains transmission values. Important: if not provided, but strings 
+        contains transmission values. Must be provided if instru_res contains
+        strings (filter filenames). Important: if not provided, but strings 
         are detected in instru_res, the default file reader will be used. 
         It assumes the following format for the files:
             
@@ -628,8 +638,8 @@ def resample_model(lbda_obs, lbda_mod, spec_mod, dlbda_obs=None,
         
     Returns
     -------
-    lbda_obs, spec_mod_res: 2x 1d numpy array
-        Observed lambdas, and resampled model spectrum (at those lambdas)
+    lbda_obs, spec_mod_res: 2d numpy array
[…]
