### 2022-07-27 ac31fa9
Optimized convolution+resampling procedure for faster handling of very high specctral resolution model to be resampled
