### 2022-07-29 #4456 @jgagneastro
> The answer is yes: the main routines of `special` are in fact compatible with the input spectrum of any kind of object regardless of the way the spectrum was obtained (provided a relevant model grid or template library), although a number of tools are indeed specific to directly imaged companions (e.g. calculation of spectral correlation in an IFS datacube).

This is great, I think it would be worth mentioning this in the paper & docs !
