### 2022-07-29 #4456 @jgagneastro
> You're right this is risky (I do not know either whether these were tested at low spectral resolution). At the moment, a warning is raised when the routines calculating spectral or gravity indices are called with a low-resolution input spectrum. I can adapt the behaviour to raise an error instead, when the wavelength sampling of the input spectrum is sparser than a given threshold, if you deem it more appropriate (and remove the example from the tutorial). What do you think?

I think an appropriate warning is fine. Say something like "these spectral indices have not been tested or calibrated at such a low spectral resolution (see Allers et al. 2007 for details)."
