### 2022-07-29 #4456 @jgagneastro
I have one issue when trying to follow the example analysis with special. It seems like my pip install did not download the special-extras directory. Is this intended ?
