### 2022-07-29 #4456 @jgagneastro
@VChristiaens In Line 29 of the tutorial (https://special.readthedocs.io/en/latest/tutorials/walkthrough.html#), I was unable to run:

`from utils import filter_reader`
