### 2022-07-29 #4456 @jgagneastro
I would define unit_obs='si' inside of the block [38] in the tutorials since this is the first place where it is used
