### 2022-07-29 #4456 @jgagneastro
I get an error when launching line [38] in the tutorial:

```
res = mcmc_spec_sampling(lbda, spec, spec_err, d_st, dlbda_obs=dlbda, units_obs=units_obs,
                         **instru_params, **mcmc_params, **conv_params, **model_params)
```

ModuleNotFoundError: No module named 'utils'

I have copied tutorial/utils.py in my main special directory to be sure it would be detected by python. I can do "import utils" without issues but the MCMC still seems unable to find it.
