### 2022-07-29 #4456 @jgagneastro
Another irregularity: When I run the tutorial as explained in the docs (nproc=1), I get only a 2-parameters MCMC output with parameters Rbb1 and Tbb1.
