### 2022-08-02 1de896e
Updated warning when trying to calculate spectral indices with low spectral resolution spectra
