### 2022-08-02 1274e86
Clarified the fact that the main routines in special are versatile (paper, github and docs), and usable on any input spectrum
