### 2022-08-04 f628f3f
Added strong warning for attempts of spectral indices calculation on very low resolution spectra
