### 2022-08-05 #4456 @VChristiaens
> I have one issue when trying to follow the example analysis with special. It seems like my pip install did not download the special-extras directory. Is this intended ?
>
> EDIT: @VChristiaens I see that the special-extras package weighs ~30 MB and that it is stored separately on your GitHub, so I suspect this was intended. However, this should be said more explicitly in the tutorial section. Otherwise, it seems like the example code should work as is after a simple installation of special.

Indeed, the tutorial is meant to be run from either downloading/cloning the special-extra directory, or running it on Binder. I have now added a line at the top of the tutorial with that disclaimer. As you guessed, the size of data+models+templates associated to the tutorial is such that it would slow down any import of special if included in the main package, hence the need for a special-extra directory.
