### 2022-08-05 #4456 @VChristiaens
> I would define unit_obs='si' inside of the block [38] in the tutorials since this is the first place where it is used

This is now added above cell [38] as recommended.
