### 2022-08-05 #4456 @VChristiaens
> Another irregularity: When I run the tutorial as explained in the docs (nproc=1), I get only a 2-parameters MCMC output with parameters Rbb1 and Tbb1.

If you mean the MCMC in Section 4.2, this is the intended output (blackbody fit).
