### 2022-10-05 #4456 @jgagneastro
Sorry to be so slow to respond Dan, it totally slipped. I'm on the verge of leaving for a paternity leave but I'm getting back on this now

Jonathan Gagné, Ph.D.
Conseiller Scientifique / Scientific Advisor
Planétarium Rio Tinto Alcan | Espace pour la vie
Professeur associé / Adjunct professor, Université de Montréal
4801 av. Pierre-de Coubertin, Montréal, Québec H1V 3V4
***@***.***

> On Oct 4, 2022, at 7:13 AM, Dan Foreman-Mackey ***@***.***> wrote:
>
>
> @VChristiaens <https://github.com/VChristiaens> — Just an update that I've contacted @jgagneastro <https://github.com/jgagneastro> a number of times also via email over the past few weeks and I'm hoping for a response, since I'd much rather not assign a new reviewer. I'll keep you updated!
>
