### 2022-10-05 #4456 @jgagneastro
> > I get an error when launching line [38] in the tutorial:
> > ```
> > res = mcmc_spec_sampling(lbda, spec, spec_err, d_st, dlbda_obs=dlbda, units_obs=units_obs,
> >                          **instru_params, **mcmc_params, **conv_params, **model_params)
> > ```
> >
> >
> >
> >
> >
> >
> >
> >
> >
> >
> >
> > ModuleNotFoundError: No module named 'utils'
> > I have copied tutorial/utils.py in my main special directory to be sure it would be detected by python. I can do "import utils" without issues but the MCMC still seems unable to find it.
> > Edit: Running MCMC with nproc=1 worked as intended (but this should still be fixed)
>
> I'm a bit puzzled regarding how to solve this one as I can't reproduce the issue. I've slightly adapted the multiprocessing wrapper contained in the `config.py` file of `special` - could you try again and let me know if it works? If it still doesn't work, could you provide the full error message to help me identify the root of the issue?

Ok I'm unable to reproduce this error now (I had to reinstall since I moved on a new laptop since last time) - either I had a problematic configuration, or the change you made fixed the problem for me. I'm now at the point where the MCMC is successfully running (still ongoing).
