### 2022-10-06 #4456 @jgagneastro
Okay, I'm done running the walkthrough, and everything ran correctly. I'm all good with the package now, and I'll check the last boxes. Thanks for your patience !
