### 2022-10-08 531063c
Adding journal name macros
