+++ paper.bib
@@ -1,3 +1,67 @@
+% Expand NASA ADS journal macros
+% ref: https://ui.adsabs.harvard.edu/help/actions/journal-macros
+@string{aas="American Astronomical Society Meeting Abstracts"}
+@string{aj="Astronomical Journal"}
+@string{actaa="Acta Astronomica"}
+@string{araa="Annual Review of Astron and Astrophysis"}
+@string{apj="Astrophysical Journal"}
+@string{apjl="Astrophysical Journal, Letters"}
+@string{apjs="Astrophysical Journal, Supplement"}
+@string{ao="Applied Optics"}
+@string{apss="Astrophysics and Space Science"}
+@string{aap="Astronomy and Astrophysics"}
+@string{aapr="Astronomy and Astrophysics Reviews"}
+@string{aaps="Astronomy and Astrophysics, Supplement"}
+@string{aplett="Astrophysics Letters"}
+@string{apspr="Astrophysics Space Physics Research"}
+@string{azh="Astronomicheskii Zhurnal"}
+@string{baas="Bulletin of the AAS"}
+@string{bain="Bulletin Astronomical Institute of the Netherlands"}
+@string{caa="Chinese Astronomy and Astrophysics"}
+@string{cjaa="Chinese Journal of Astronomy and Astrophysics"}
+@string{dps="American Astronomical Society/Division for Planetary Sciences Meeting Abstracts"}
+@string{fcp="Fundamental Cosmic Physics"}
+@string{gca="Geochimica Cosmochimica Acta"}
+@string{grl="Geophysics Research Letters"}
+@string{iaucirc="IAU Cirulars"}
+@string{icarus="Icarus"}
+@string{jaavso="Journal of the American Association of Variable Star Observers"}
+@string{jcap="Journal of Cosmology and Astroparticle Physics"}
+@string{jcp="Journal of Chemical Physics"}
+@string{jgr="Journal of Geophysics Research"}
+@string{jqsrt="Journal of Quantitiative Spectroscopy and Radiative Transfer"}
+@string{jrasc="Journal of the RAS of Canada"}
+@string{maps="Meteoritics and Planetary Science"}
+@string{memras="Memoirs of the RAS"}
+@string{memsai="Mem. Societa Astronomica Italiana"}
+@string{mnras="Monthly Notices of the RAS"}
+@string{na="New Astronomy"}
+@string{nar="New Astronomy Review"}
+@string{nat="Nature"}
+@string{nphysa="Nuclear Physics A"}
+@string{pasa="Publications of the Astron. Soc. of Australia"}
+@string{pasp="Publications of the ASP"}
+@string{pasj="Publications of the ASJ"}
+@string{physrep="Physics Reports"}
+@string{physscr="Physica Scripta"}
+@string{planss="Planetary Space Science"}
+@string{pra="Physical Review A: General Physics"}
+@string{prb="Physical Review B: Solid State"}
+@string{prc="Physical Review C"}
+@string{prd="Physical Review D"}
+@string{pre="Physical Review E"}
+@string{prl="Physical Review Letters"}
+@string{procspie="Proceedings of the SPIE"}
+@string{psj="Planetary Science Journal"}
+@string{qjras="Quarterly Journal of the RAS"}
+@string{rmxaa="Revista Mexicana de Astronomia y Astrofisica"}
+@string{skytel="Sky and Telescope"}
+@string{solphys="Solar Physics"}
+@string{sovast="Soviet Astronomy"}
+@string{ssr="Space Science Reviews"}
+@string{zap="Zeitschrift fuer Astrophysik"}
+
+
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
