### 2022-10-08 dfd13a2
Updating journal names
