### 2022-10-08 #4456 @dfm
@VChristiaens — I've opened a PR with some edits and questions. Please address those and then go through the following steps:

1. Take one last read through the manuscript to make sure that you're happy with it (it's harder to make changes later!), especially your name and affiliation. I've taken a pass and it looks good to me!
