### 2022-10-10 e1a10dc
Clarification of Eq. 1
