+++ paper.md
@@ -99,15 +99,17 @@ instruments (e.g. photometry+spectroscopy):
 
 \begin{equation}
 \label{Eq:logL}
-\log \mathcal{L}(D|M) = - \frac{1}{2} \big[\mathbf{W}(\mathbf{F_{\rm obs}}-\mathbf{F_{\rm mod}})^T\big] \mathbf{C^{-1}} \big[\mathbf{W}^T(\mathbf{F_{\rm obs}}-\mathbf{F_{\rm mod}})\big]
+\log \mathcal{L}(D|M) = - \frac{1}{2} \big[\mathbf{W}\odot(\mathbf{F_{\rm obs}}-\mathbf{F_{\rm mod}})\big]^T \mathbf{C^{-1}} \big[\mathbf{W}\odot(\mathbf{F_{\rm obs}}-\mathbf{F_{\rm mod}})\big]
 \end{equation}
 
-where $\mathbf{F_{\rm obs}}$ and $\mathbf{F_{\rm mod}}$ are the fluxes of the 
-observed and model spectra respectively, $\mathbf{C}$ is the spectral 
-covariance matrix, and $\mathbf{W}$ is the vector of weights 
-$w_i \propto \Delta\lambda_i/\lambda_i$, with $\Delta\lambda_i$ the width of 
-spectral channels (for integral field spectrograph points) or the FWHM of 
-photometric filters.
+where $D$ are the data at hand (measured fluxes and spectral covariance), 
+$M$ is the considered model, $\mathbf{F_{\rm obs}}$ and $\mathbf{F_{\rm mod}}$ 
+are the fluxes of the observed and model spectra respectively (each are vectors of 
+length $n_z$, the number of spectro-/photometric points), $\mathbf{C}$ is 
+the spectral covariance matrix ($n_z$x$n_z$), $\odot$ stands for the Hadamard product, and 
+$\mathbf{W}$ is the vector of weights $w_i \propto \Delta\lambda_i/\lambda_i$ (length $n_z$), 
+with $\Delta\lambda_i$ the width of spectral channels (for integral field spectrograph 
+points) or the FWHM of photometric filters.
 
 A Jupyter notebook tutorial illustrates most available features in ``special`` 
 through their application for the analysis of the composite spectrum of CrA-9 B/b 
