### 2022-10-10 b51d853
incremented version to reflect minor changes requested during peer-review process
