### 2022-10-10 #4456 @VChristiaens
Thanks for the PR. I have now updated the draft after adapting Eq. 1 and clarifying each of its components as recommended. I also went through steps 1-3 above:
- the incremented version number is 0.1.4.
- this is the DOI of the release: 10.5281/zenodo.7181950

Thanks again for leading the review process @dfm, and thanks @QuentChangeat and @jgagneastro for thoroughly testing the package and providing feedback!
