+++ paper.bib
@@ -61,7 +61,7 @@
 	date-modified = {2022-05-19 16:30:29 +0200},
 	doi = {10.1086/501068},
 	eprint = {astro-ph/0508461},
-	journal = {Astrophysical Journall},
+	journal = {Astrophysical Journal Leters},
 	keywords = {Cosmology: Theory, Astrophysics},
 	month = feb,
 	number = {2},
