### 2022-10-10 4818880
Minor typo fix in paper.bib
